Implement UserService (#30)
* Start User model...

* Add forename and surname to User.php

* Add email to User

* Add role to user and add docs

* Add create User to UserRepository

* Add edit to User

* Add deleteUser to UserRepository

* Add FindUserByEmail to UserRepository

* Add findUserBySurname() to UserRepository

* Add findAllUsers() to UserRepository

* Add password to User

* Add UserInterface and return types

* Add createUser to UserInterface

* Add registerUser() to UserInterface

* Add authUser() to UserService

* Add role class, roleName() and roleStatus()

* Add approve() and disapprove() to role

* Add approve() and disapprove() to Userservice

* Fix Role because law of demeter

* Add registerTrainee() and registerTrainer to UserService

* Add disapprove() and approve() to UserService

* Add identifier to user

* Add findUserById() to UserRepository

* Rename findUserbyEmai/Surname to findUserByEmail/Surname

* Rename UserInterface and UserRepository

Rename UserInterface to UserRepository and
UserRepository to UserInMemoryRepository

* Add Exception to UserRepository

UserRepository throws exception if email is already in use

* Add password policy to UserInMemoryRepository

* Fix view/read approved reports

* Add status revised to report

* Add status revised to view

* Remove role() from User.php and Views/User.php

* Refactoring UserTest and deleted RoleTest

* Create UserFileRepository and add createUser()

* Add deleteUser() to UserFileRepository

* Add findAllUsers() to UserFileRepository

* Add findUserByEmail() to UserFileRepository

* Add findUserBySurname() to UserFileRepository

* Add findUserById() to UserFileRepository

* Fix doc comments

* Fix doc comments and add exceptions

* Add milestones to README.md

+++ app/views/TraineeView.php
@@ -8,25 +8,33 @@
         <th>Aktionen</th>
     </tr>
     <?php foreach ($this->reports as $report):
-         $reportId = $report->id(); ?>
+         $reportId = $report->id();
+         $traineeId = $report->traineeId(); ?>
         <tr>
             <td><?php echo substr($report->content(), 0, 10); ?></td>
             <td><?php echo $report->date(); ?></td>
             <td><?php echo $report->calendarWeek(); ?></td>
             <td><?php echo $report->status(); ?></td>
-            <?php if ($report->status() === Report::STATUS_NEW
-                || $report->status() === Report::STATUS_DISAPPROVED
-                || $report->status() === Report::STATUS_EDITED): ?>
                 <td>
                     <ul>
-                        <li><a href="/report/editReport?reportId=<?php echo $reportId; ?>">Bearbeiten</a></li>
-                        <?php if ($report->status() !== Report::STATUS_DISAPPROVED): ?>
+                        <?php if ($report->status() !== Report::STATUS_APPROVED
+                                && $report->status() !== Report::STATUS_APPROVAL_REQUESTED): ?>
+                            <li><a href="/report/editReport?reportId=<?php echo $reportId; ?>">Bearbeiten</a></li>
+                        <?php endif; ?>
+                        <?php if ($report->status() !== Report::STATUS_DISAPPROVED
+                                && $report->status() !== Report::STATUS_APPROVED
+                                && $report->status() !== Report::STATUS_APPROVAL_REQUESTED): ?>
                             <li><a href="/report/requestApproval?reportId=<?php echo $reportId; ?>&action=requestApproval" onclick="return confirm('Soll der Bericht eingereicht werden?')">Einreichen</a></li>
-                            <li><a href="/report/delete?reportId=<?php echo $reportId; ?>&action=delete" onclick="return confirm('Soll der Bericht wirklich gelöscht werden?')">Löschen</a></li>
+                            <?php if ($report->status() !== Report::STATUS_REVISED): ?>
+                                <li><a href="/report/delete?reportId=<?php echo $reportId; ?>&action=delete" onclick="return confirm('Soll der Bericht wirklich gelöscht werden?')">Löschen</a></li>
+                            <?php endif ?>
+                        <?php endif; ?>
+                        <?php if ($report->status() === Report::STATUS_APPROVED
+                                || $report->status() === Report::STATUS_APPROVAL_REQUESTED): ?>
+                            <li><a href="/report/viewReport?reportId=<?php echo $reportId; ?>&traineeId=<?php echo $traineeId; ?>">Öffnen</a></li>
                         <?php endif; ?>
                     </ul>
                 </td>
-            <?php endif; ?>
         </tr>
     <?php endforeach; ?>
 </table>
